Move UserAvailability into a dedicated availability app
Relocate the existing weekly-availability feature out of accounts and
home into a new `availability` app, with no behavior change.
