+++ accounts/admin.py
@@ -17,7 +17,6 @@ from accounts.models import (
     ButtondownAccount,
     CustomUser,
     Link,
-    UserAvailability,
     UserProfile,
 )
 from home.models.session import SessionMembership
@@ -181,31 +180,6 @@ class ButtondownAccountAdmin(DescriptiveSearchMixin, admin.ModelAdmin):
     search_fields = ["buttondown_identifier", "user__email"]
     list_display = ["user", "buttondown_identifier"]
     list_filter = (RelatedUserPastDjangonautFilter, RelatedUserPastSessionMemberFilter)
-
-
-@admin.register(UserAvailability)
-class UserAvailabilityAdmin(DescriptiveSearchMixin, admin.ModelAdmin):
-    """Admin interface for UserAvailability."""
-
-    list_display = ("user", "slot_count", "updated_at")
-    search_fields = (
-        "user__username",
-        "user__email",
-        "user__first_name",
-        "user__last_name",
-    )
-    list_filter = (
-        RelatedUserPastDjangonautFilter,
-        RelatedUserPastSessionMemberFilter,
-        "updated_at",
-    )
-    readonly_fields = ("updated_at",)
-    raw_id_fields = ("user",)
-
-    @admin.display(description="Number of Slots")
-    def slot_count(self, obj: UserAvailability) -> int:
-        """Display the number of availability slots."""
-        return len(obj.slots)
 
 
 class UserModelInline(admin.TabularInline):
