+++ accounts/factories.py
@@ -1,7 +1,7 @@
 import factory
 from django.db.models.signals import post_save
 
-from accounts.models import CustomUser, UserAvailability, UserProfile
+from accounts.models import CustomUser, UserProfile
 
 
 @factory.django.mute_signals(post_save)
@@ -23,11 +23,3 @@ class UserFactory(factory.django.DjangoModelFactory):
     last_name = "Doe"
     email = factory.LazyAttribute(lambda obj: f"{obj.username}@example.com")
     profile = factory.RelatedFactory(ProfileFactory, factory_related_name="user")
-
-
-class UserAvailabilityFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = UserAvailability
-
-    user = factory.SubFactory(UserFactory)
-    slots = []
