+++ accounts/urls.py
@@ -8,7 +8,6 @@ from .views import DeleteAccountView
 from .views import profile
 from .views import ResendConfirmationEmailView
 from .views import SignUpView
-from .views import UpdateAvailabilityView
 from .views import UpdateEmailSubscriptionView
 from .views import UpdateUserView
 
@@ -41,11 +40,6 @@ urlpatterns = [
         "email_subscriptions/",
         UpdateEmailSubscriptionView.as_view(),
         name="email_subscriptions",
-    ),
-    path(
-        "availability/",
-        UpdateAvailabilityView.as_view(),
-        name="availability",
     ),
     path(
         "delete/",
