+++ accounts/views.py
@@ -31,8 +31,6 @@ from .forms import CustomUserCreationForm
 from .forms import DeleteAccountForm
 from .forms import EmailSubscriptionsChangeForm
 from .forms import INTERESTED_IN_FIELDS
-from .forms import UserAvailabilityForm
-from .models import UserAvailability
 from .tasks import delete_user_account
 from .tokens import account_activation_token
 
@@ -237,28 +235,6 @@ class UpdateEmailSubscriptionView(LoginRequiredMixin, UpdateView):
             self.request,
             messages.INFO,
             "Your profile information has been updated successfully.",
-        )
-        return reverse("profile")
-
-
-class UpdateAvailabilityView(LoginRequiredMixin, UpdateView):
-    """View for updating user's weekly availability."""
-
-    form_class = UserAvailabilityForm
-    template_name = "registration/update_availability.html"
-
-    def get_object(self, queryset=None):
-        """Get or create the UserAvailability object for the current user."""
-        availability, created = UserAvailability.objects.get_or_create(
-            user=self.request.user
-        )
-        return availability
-
-    def get_success_url(self):
-        messages.add_message(
-            self.request,
-            messages.SUCCESS,
-            "Your availability has been updated successfully.",
         )
         return reverse("profile")
