+++ availability/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AvailabilityConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "availability"
+    verbose_name = "Availability"
