+++ home/models/survey.py
@@ -7,7 +7,7 @@ from django.utils.safestring import mark_safe
 from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 
-from accounts.models import UserAvailability
+from availability.models import UserAvailability
 from home import email
 from home.managers import (
     SurveyQuerySet,
