+++ indymeet/urls.py
@@ -16,6 +16,7 @@ urlpatterns = [
     path("admin/", include(wagtailadmin_urls)),
     path("documents/", include(wagtaildocs_urls)),
     path("accounts/", include("accounts.urls"), name="accounts"),
+    path("accounts/", include("availability.urls")),
     path("", include("home.urls")),
     path("", include("puput.urls")),
 ]
